+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # CHANGELOG
 
+# v2.1.0
+
+* Include the response code and body in `HttpError#to_s`
+
 # v2.0.0
 
 * `TokenProvider` added